#78 implemented testing for user db
Tests to make sure the user cannot be created without a name, uid, provider, account_type, or email. (5 different tests)

Made it so that user requires all 5 of those.

Fixed the facebook login problem (was a deployment issue on facebook)

Commented a lot of controller test default code.... It was throwing a million errors.